backend/requirements.txt
asgiref==3.8.1
black==25.1.0
boto3==1.38.27
botocore==1.38.27
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.0
colorama==0.4.6
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jmespath==1.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
